Load command requests cache retrieval
