+++ EssentialFeed/Feed Cache/FeedStore.swift
@@ -14,5 +14,6 @@ public protocol FeedStore {
     
     func deleteCachedFeed(completion: @escaping DeletionCompletion)
     func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion)
+    func retrieve()
     
 }
