+++ EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -36,6 +36,10 @@ public final class LocalFeedLoader {
             completion(error)
         }
     }
+    
+    public func load() {
+        store.retrieve()
+    }
 }
 
 private extension Array where Element == FeedImage {
